+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.1.8] - 2024-04-01
+### Added
+- Support for additional imports and compiler options of themes.
+
+### Changed
+- Sort themes alphabetically.
+
 ## [0.1.7] - 2024-04-01
 ### Added
 - Support for unstable flags.
@@ -38,6 +45,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2024-03-22
 First version
 
+[0.1.8]: https://github.com/lumeland/init/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/lumeland/init/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/lumeland/init/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/lumeland/init/compare/v0.1.4...v0.1.5